:pencil: Add comments
In the base case, we only need to determine whether t.left <= t.node <= t.right
Similarly, we can combine lower trees into a single value IF those trees are valid.

The corner-case is when the Tree is valid, but one of the values is invalid in the
greater Tree. For example given this Tree:
    3
   / \
  2   4
 / \
1   6

We can quickly see that while 1 <- 2 -> 6 is a valid tree, the
6 is not valid in the greater Tree. In this case we need to compare
the larger value with The 3. This is the opposite on the right-hand
side of the root node

Therefore, in the case of a Tree on the left side of the root, we pass
in the 'l' flag to the function and return the t.right value. In the
'r' case we do similar with the t.left value.