:zap: Improved performance
This approach uses sets to determine which values can be chosen from
and the random.choice() function to choose a random value in an iterable

i.e. {1, 3, 5} - {1, 5} = {3}

Create a list across the range from 0 to n - 1 and remove the elements
in l. Then choose a random value in that new list