:books: Learn something today regarding anonymous functions
Not sure if either of my solutions actually solve the problem, but
interesting nevertheless.

In this case, the issue stems from the fact that  there is no value passed to
the function and so it is called with CURRENT value of i. That is set to 9.
Therefore the loop will print '9' ten times

Anonymous functions are very interesting. This isn't what I'd expect on first
glance.