:sparkles: Solve for the largest path that includes a node
The algorithm implemented works as follows:
1. In the simple case, determine which is the largest
   At each level, determine which is the largest that could end here
   node + left, node + right or 0
2. Determine whether that is larger than the current largest sum
3. At a greater level, step into each of the nodes and determine
   the same outcomes. What is the greatest value that ends here,
   and is it greater than the current greatest value
4. As a final check, detemine that the root node isn't greater than
   any of the other options

Note: The 0 check is important here. If at any point the max_ending_here
      values falls below 0, that path is finished and the max_ending here
      is set to 0.

Time complexity here is O(n + l) when n & l are the number of nodes + leaves
respectively; as each value only gets visited once. or O(V) where V is the
number of values in the graph.