:sparkles: Return next greatest permutation
The idea here is to create a generator that produces permutations
as it is called. Provided that the function is provided with a sorted
list to produce permutations of, the resulting generator will be sorted.

Therefore, to return the next iteration of the generator, once the
generator has matched the input list, the next value of the generator
is returned manually with the __next__() call.

In the case that the provided value is already the greatest, the generator
will raise a StopIteration error. In this case, we return the sorted input
list.

Note that sorting the list will take NlogN, assuming a very large input list,
the sorting may take a while. The generator thankfully will not use up any memory in
the case of large input lists.

Note that at the cost of allocating extra memory, the input list can be sorted only
Once instead of potentially twice.