:sparkles: Map implementation
The idea here is to use a default dict to save some LOC and improve
readibility. From there we save the last time value pair and dynamically
create all values up to the new value when we add a new time, value pair
to the map.

The get function works by checking if the value that we are looking for
exists in the map. If not, if it's greater than the last value then we simply
return the final set value. If it's less that this value and it's not in
the map, then it has not been set yet, and we return None.

There is a low-hanging improvement that can be made here - to dynamically
determine whether the value to get is within the range of the previous
several (or more) values. This becomes more difficult to understand and
we suffer readibility.