:sparkles: Modified DFS - should have seen this earler
Effectively a relatively simple depth first search algo.
The steps that can be taken are x-1, y; x+1, y; x, y-1 and x, y+1.
Therefore a vistited list is created such that we don't recursiveley
keep searching the same dead ends.

DFS is slightly more appropriate for this use case than bfs, and should
be a lot faster.

The concern that the same letter may not be used more than once is handled
succinctly by the visited list, as a byproduct of DFS.