:sparkles: Find the length of the longest consecutive elements sequence
The idea here is to use a set and determine whether the value before
or after the value that we are looking at is in the set.
This is called recursively and calls both up and down at the same time.

Converting the list to a set would take 0(N) - One option here is to
add items to the set as we iterate through the list. That would require
some refactoring.