:zap: Improve performance
Remove the O(N) conversion of list to set. This iteration appends
items to the set as they come up in the iterable.

Unfortunatley, this can still have a relatively poor worst case
time complexity: In the case of [1, 2, 3, 4, 5] the worst case is
O(N^2) as for each iteration it will need to run through all the
values each iteration. In that case, the previous option would be
faster.

One potential performance improvement here would be to save the
depth of each range for each number. i.e. for 1 the depth = 1,
for 2, the depth is 2, and the upper and lower bounds are 1 and 2
respectively. For 4, the length is 4, upper is 4 and lower is 1.
This is relatively memory intensive, for example in the case where
we have [1, 3, 5, 7, 9] we have 5 unique 1 depth long elements.

As often, there is a tradeoff between Time and Space complexity