:sparkles: Count minimum number of steps on an infinite board
To start we place all the points on a queue.

For the simple case we simply determine if the next point is
in the range of the current point. This takes at most 8 iterations
per check. Another way of doing this is to generate a set of all
points that are in range and check if the next point is in that.
That would be slower as we would need to calculate all 8 points
regardless of where it is.

If this point is in range, we take the current point off of the
queue.

If the next point can not be reached in a single step, then the
distance to that next point is minimized by adding another point
to the queue that is between the current point and the next point.

This logic is handled by the minimize distance function which takes
8 iterations each and every time.

When there is only a single point left on the queue, the total number
of steps is returned. This works because there are two paths from the
conditional, either remove an element from the queue, or add an element
to the queue. Therefore, if there is no path from the penultimate element
to the last element, another element is added to the queue. Then there
are two elements in the queue and the loop does not terminate.