:sparkles: Return a contiguous subset of a list that adds up to K
Base cases: If the entire subset is equal to K or K is in the list

Then create a slice of n - 1 elements and determine whether those
two options are equal to K. Then reduce the size of the slice
and slide it through the list.

i.e. For [1, 2, 3, 4, 5] we consider [1, 2, 3, 4] and [2, 3, 4, 5]
For the next step, we consider [1, 2, 3], [2, 3, 4] & [3, 4, 5]
etc until the steps are of size 2.