:sparkles: Calculate whether a LinkedList is a palindrome
To be honest this isn't a very good solution. I would have liked
to have a queue (or dequeue) that I could pop elements off of
as we iterated through the list. But by not knowing the size of
the LL it's difficult to know whether we are at the midpoint or not.

In the case of 1 -> 2 -> 2 -> 1 -> 1 -> 1 Even though this is not
valid, the queue pop approach would say that it is.

The solution here involves making a list of all elements and determining
whether the 2nd half is a mirron image of the first half.

More efficient data scructures may be able to save a few operations, but nothing
too special.

A doubly linked list MAY allow us to iterate from both sides at the same time,
but I'm not sure of that