:sparkles: Simple solution might be able to optimize
Simply hop from one position to the next, refilling hops as you
go along. Question seems to infer that from that point one resets
their hops to the value at that value. If that is the case this
solution is accurate, else we can easily adapt it to fit the other
case