:zap: Improve performance using backtracking and recursion
Using backtracking and recursion, determine whether there are enough
blocks to get to the last point.

For the sake of an example: consider [a, b, c, x, y, z, 0]. To get to
the last point, we need either,a, = 6, b = , c = 4 x = 3, y = 2
or z = 1. Assuming that x is 3, we similarly require
a = 3, b = 2 and c = 1. This then becomes a recursive problem:
Solve for 0, x = 3, solve for x, a = 3.

Since there are many paths to take, without calculating the possibilities
of what values would be required, we can use backtracking to eliminate
all paths that do not provide us with a solution.