:sparkles: Implement can_shift algorithm
Logic here is to check substrings of A are in B.
The worst case time complexity is relatively poor here to be
honest.

i.e. for A = "abcde" and B = "cdeab" we check that "bcde" is in
B and "a" is the rest of the string. in the next iteration, we
check that "cde" is in the string, and that "ab" is the first part
of the string. Since this is the case, we can return True