:zap: Improve solution to use one line.
Set a bit mask of all even numbers (170) and all odd numbers (85)

Bitwise & the input number with the even bitmask and bitshift one
position to the right.

Bitwise & the input with the odd bitmask and bitshift one position
to the right.

Bitwise OR the two results together. Not very readable but works
for 8 bit numbers just fine.