:zap: Improve performance
New algorithm is 300 times quicker over 1000 iterations.

This appoach runs the matching in O(N), with a single sort
per i in range, this could potentially be slow, and could
still be improved be keeping the previously sorted string,
this seems like an unneccessary optimization and we will
pay for it in readability.
Some results:
  Initial algo: Time taken = 3.3350119590759277 seconds
  Faster algo: Time takes = 0.017987728118896484 Seconds