:sparkles: Another Tree problem
Solved this by recursively checking whether any of the subtrees
are equal to the other tree. This piggy backs nicely off the
__eq__ method for the Tree class that I created a while back.

A type check was added to the __eq__ method to improve the
comparison and remove some errors. There is no reason to check
whether a Tree is equal to an integer.