:sparkles: Create a generator that returns a Tree
This should create trees as the generator is called. The
arbitary large is handled by the > 0.9 step. This means that
we should generate multiple trees before stopping. This
solution chooses either the left or right branch to continue
down, it could be called recursively to do both.

Without having the interviewer to speak with, it's tough to
see what they want, but this is 0(1) and 'large'