:sparkles: Some sort of queue solution
Check the minimum value of the root node. Then add all lefts
and rights for each node in the queue to a new queue.
At the same time, calculate the total for that layer and
store it in a dictionary.

Finally return the layer with the smallest number.

Easy fixes to only keep the lowest number