:sparkles: Use two pointers to sort squares
Approach here is to use two pointers, one at the head and
one at the tail of the queue. Since we basically have two
sorted lists, we simply need to zip them together.
This is done in a single list with a pointer at the head and
the tail instead of two lists. The rest is very similar.

Time complexity O(N) since we need to check each element at least
once. Does use additional space of O(N).