:sparkles: Check for the smallest set in a range of intervals
Loop through the numbers, and add the ones that have not been seen
before to a set. If we see another number that matches the one that
we have already seen, then we confirm that number by adding it to
a set of confirmed numbers.

Any sets that are not confirmed are added to a list of unconfirmed values.
If there is no value in the confirmed values that is in the interval,
then we add a number from that interval to the set of values.

Potentially not going to work and has worse case of O(N*N) when N is the
number of elements.