:zap: Improve performance
Instead of using the incredibly slow recursive solution, we
can use a slightly faster recursive solution.

Basically, one way to determine a palindrome is to head from the
end and the front at the same time, and check that the chars all
match. In the case that one of the chars does not match, we
recursively call the function again with the string without the
leading or trailing character. Since we are only allowed k
changes, this is a backtracking algo that returns true if a
solution is found.

We return True once there is only one char remaining, or we return
True once the head index and the tail index are the same.