:sparkles: Really neat recursive solution
Quite happy with this approach, the greedy method was
going to fail so I chose a "dfs" analogue that returns the
max between each point, recursively until it reaches the end of
the coin list. The time complexity here is not very good at all
though. Since we do really need to consider every block, it
might be difficult to speed this up too much..