:zap: Performance improvements
Use a binary search approach but also get the upper and lower
bounds by doubling the bounds each time. This makes it quck
for very large numbers. Then a simple binary search to return
the value. Since floating points are ugly to work with, the
epsilon here is chosed to 0.001 and then the midpoint is returned.
This is a good approximation.