:sparkles: Get the next values in an inoreder traversal
Idea here is to find the element that you are concerned with,
and then find the smallest element in the right subtree.

Finding the smallest element of the subtree is done by traversing
down the left hand side while it exists.

The final concern is when there is no right subtree. In that case
we use the parent - and do the same thing.