:sparkles: Difficult problem, BFS with limits to solve
Took a few different paths to try and solve this one.
Ended up with perhaps not the neatest but an accurate
solution to the problem.

Algorthithm works like this:
1. Find a '1' in the board.
2. Determine the rectangle's limits by heading right and down
   until a 0 is reached
3. Using those limits as the bounds to a BFS (that can only)
   head right and down, count the number of 1s

Some obvious speedups won't work. We can't mark blocks as
visited in the greater board because different shapes of
rectangles can us the same blocks.

DFS wasn't an obvious pattern here either, with the limits
and the fact that we needed to check all the blocks