:art: Performance and structure improvements :zap:f
Rather than implement the copy of the iterator, we can
keep track of the next element in a list (can use an int also).

This allows us to return that value if next is called, and keep
returning it so long as peek() is called. A similar flow is
implemented for hasNext().