:sparkles: Implemnent 3 stacks using one list
Potentially uses a bit of extra space, but it's not
as bad as it could be. Probably also a bit slow to
lookup some values.

The idea here is to extend the stack with three items per
push. Obviously if we don't need to extend, we don't, else
we change a value from None to the value that it needs to
be. In the instance of removing values, if all three
of the last values are None, then those are removed.

If there is no value in the last three, we recursively
check the stack for a value.

There is a potential issue with filling up from the front
and not the back. It's faster to lookup from the front,
and uses the same amount of memory (as those Nones are
still allocated). Therefore this is a better approach.