:sparkles: Wildcard matching balanced parentheses
Implement a backtracking algorithm, for each wildcard,
recursively check whether the resulting pattern solves
for '(', ')' or ''.

Also check for obvious isssues such as the first element being
). This is handled by the b < 0, and the last element being (
which is handled by the return not b.

A faster implementation where we predict the value to be popped
off the list is not feasible as that result may be handled later
in the string. There may be a small performance increase in
guessing a value "better" but that would require a bit more
space and would become harder to understand.