:sparkles: Recursively use the base 2 gray codes
The first set of grey codes we can prepend a 0 to.
The second set of grey codes we prepend a 1 to. Since we
can only change one bit per iteration, we reverse the
second set of numbers so that the only number that is
different is the prepended 0 -> 1. Then since the
codes only change once per cycle on the way up, they
will also only change once per cycle in the reverse
direction.