:zap: Improve performance
This is almost 1000 times quicker over a 500k entry sum.

This approach works much better for similar lookups. In the
instance that we already have a similar lookup, this approach
will rather use that value and add the rest of the numbers to
that number. This is only used when it's quicker to do the
lookup (x - j) than to simply add all the numbers (j - i).

This approach does have some limitations, for one when we have lots
of keys in each dictionary it may take long to find a key that is
faster than a clean lookup, but in "general" it should be more
performant.