:sparkles: Return the closest k points to cp in a list
Using the distance formula, create a dictionary of all the points
and their distances, then return k elements from the dict
sorted on distance. This approach is nice and readable.

Time complexity is N for the distances and then NlogN for the
sorting, so that's NlogN. We can do better.