:sparkles: Caclulate the mode, but easier
Calculate a number that exists more than half of the
time in a list. This solution uses a dictionary to
record the totals and returns when the count of one
of them is greater than half of the length of the
list.

O(N) Time complexity. Can we get O(logN)? In the example
case we get logN, but that's by chance. sorting would be
nlogn and wouldn't really help. We can't binary search
an unsorted list. Time complexity of n.count() should be
O(N) for each item, so that's N*N.