:sparkles: Find the fewest numbers to create an integer
Algorithm: Start with the sqrt of n, convert to an integer.
Note if sqrt is slow, and it might well ve, we can use a
quicker algorithm to get the an integer relatively close
with a binary search type.

Decrement that value until it's less than the value looking
for when squared, when this is the case do the same with the
next value until the value is found. Note we will always find
a solution by using 1s.