:sparkles: Check whether any permutation of a string is a palindrome
The idea here is that a pallindrome has an even number of every
letter (except for 1, in the middle). This algorithm runs through
all the letters and counts then in a dictionary. Then if there are
more than one letter that appear an odd number of times, return False.
Else return True.

Time complexity: O(len(s)) for the counting letters and O(len(d)) where
d is the number of elements in the dictionary. Worst case here of len(s)/2.
~ O(N2)