:sparkles: Slow solution here
For each word in the list, for each other word in the list,
for each letter in the word, check whether they are a match or
not. Once we have been through all the words, the maximum
number of letters that is required is appended to the wordslist

Time complexity: O(N2 + l2) where l is the lengh of the words.
Space compexity, O(N)