:zap: Improve performance
By only keeping the prefixs by editing the input in place we
improve the space complexity to O(1). The time complexity
is unchanged, but the O(l) lose meaning. it's still O(N^2)
though.

Sorting the list is O(NlogN) then we'd still need O(N) to get the
prefixs. Still think there is a better way..