:zap: Improve performance
Clever solution here, use the array as the hashmap, but use
negative numbers.

Since each number from 1 -> n should appear in the array, we can
simply negate the value in the correct position. Then if we already
have a negative value in that position, then return the value that
is a duplicate.

This is now O(N) Time complexity and O(1) Space Complexity not
including the input space.

We use the absolute value to get around the fact that some of our values
are not negative.