:art: Improve look and feel.
Recalling that the sum of N natural numbers is given by
(N*2 + N) / 2, we subtract the actual sum of the numbers o(N) from
the expected sum of the numbers. This is a O(N) and we can throw
away the input array once we have counted each number.

This works becuase we have each number from 1 - N once, and
another number twice, so we use len(l) -1 for the sum value.