:zap: Performance improvement
Perhaps not the massive performance improvement that we are looking
for, but should be vastly quicker than the previous iteration, especailly
for larger dictionaries. Ideally we would be able to lookup in 0(1)
the value from the dict, and we could make that tradeoff by having
many values that return the same value. This could also get quite
memory intensive, in which case the O(1) performance of the dict
may come apart.

For clarity, the idea is to split this into a random.choice with an
equal number of blocks of size 0.05 (20 blocks). Then each of these
is set as the key to the dictionary that allows us O(1) lookup time.

i.e. Considering 'a': (0.05, 0.1, 0.15, ..., 0.9) are all mapped to
'a' while (0.9, .. 0.975) are mapped to 'b' and (0.975, ..., 1.0)
are mapped to 'c': Then we can do a lookup in O(1). The downside of this
is space complexity (huge) and the minimum step size will need to change
as the values in the markov chain change.