:sparkles: Rotate a linkedList to the right
- Reimplment the LinkedList, to use Nodes rather
- Implement a len for the linked list
- Getting the length is O(len(l))
- The rest of the algo is also O(len(l))
- Use a modulo to handle k > n

The idea here is to find the node at k-len(l).
That node then become the end of the linked list.
(set next to None). Once this has been found, we
continue until we find the last node. The next of
the last node is set to head. Then the head is
set to the next of the k-nth node.
The order of these operations may be slightly
different that as described.