:sparkles: Integer palindrome
This approach splits the integer into two halves.
The second half is then reversed and compared with the first half.
In the case of an odd lengthed number, i.e. 121, a delta is added
so that only the values after the middle number (2 in this case) are
compared with the numbers before 2. This works simialrly for 12321
or anyting longer, obviously.

Been away for a week, good to be back.